home-manager, kakoune, repo: Move `kakoune`-config into home-manager
With this `kakoune` will be fully managed through home-manager.

Basically this means I use the `kakoune`-module from home-manager, convert the small config
I had over to it and move the package-declarations into the kakoune-block.

This way the main packages-block becomes cleaner and all the dependencies are
neatly organized and visible.

Since the path `.config/kak/kakrc` is now managed we add it to the `.gitignore`.