fontconfig, home-manager: Install fonts through home-manager
With this it is no longer necessary to install fonts system-wide.

The changes to the fontconfig were necessary because for some reason
font-selection appears to work worse this way - maybe it is just because
I'm installing much fewer fonts than before though...

This introduces two generated files in `environment.d` and `fontconfig/conf.d`
which are ignored.