home-manager, sway: Move sway configuration into home-manager
This removes the special config in `.config/sway` and replaces it by config
generated by home-manager. This also includes the themes for `sway` and `wofi`.

I think the generated config should match the old config pretty well. `home-manager`
adds some options that weren't set before but they don't look too bad.

What is definitely not working are the color-settings. I couldn't be bothered right now.

Most of the includes were removed - only the style-include remains. They weren't that
useful anyway and with `nix` I can separate different areas of concern in another way
if I want to.

I got rid of the different symlinks for themeing. Instead I now directly specify the
config/CSS I want to use in `nix` - it already manages linking files and managing
dependencies anyway.

Because of this the different theme-options moved into the `.config/home-manager` directory
so they can be included in the generated profile.

Another nice thing is that I could get rid of more explicitly installed packages. The
different components like terminal-emulator, browser or `wofi` are now referenced
by the configuration and thus detected as dependencies and included in the profile :).
 - I really like that I can keep my $PATH cleaner this way!