home-manager, elvish, zsh: Convert `zsh`-aliases to `elvish`-functions
This removes the `zsh`-alias file and replaces it with an `elvish`-module
containing functions representing these aliases.

Since `elvish` has a namespace per module, the aliases have to be declared
in `rc.elv` as well, so they are available in the global namespace.

The alias-functions now directly reference binaries from the store, allowing
`config.home.packages` to be slimmer and `$PATH` to be cleaner as well while
adding more dependency information to the home-setup :).