elvish, home-manager: Generate `direnv`-Hook through a derivation
Hopefully this means, the hook will always be up to date when upstream `direnv`
changes.

Also this make my config shorter, which is nice.

Unfortunately I couldn't get `elvish` to import the file using an absolute
path. So for now, the Hook still needs to be in `~/.config/elvish/lib` to
be picked up... :/